+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.AdEntity;
 
 @Repository
-public interface AdRepository extends JpaRepository<AdEntity, Long> {
+public interface AdRepository extends JpaRepository<AdEntity, Integer> {
 }
