+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -5,6 +5,6 @@ import ru.skypro.homework.entity.CommentEntity;
 
 import java.util.Collection;
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
+public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
     Collection<CommentEntity> findByAd_Id(long adId);
 }
